DEV-29822 Migrate code into Magento 2 format
