--- app/code/Gateway/view/frontend/layout/gateway_redirect_go.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0"?>
-<!--
-/**
- * Copyright © 2015 Magento. All rights reserved.
- * See COPYING.txt for license details.
- */
--->
-<page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" layout="1column" xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
-    <body>
-        <referenceContainer name="content">
-            <block class="IPGPAY\Gateway\Block\Redirect" name="redirect" template="IPGPAY_Gateway::redirect.phtml" />
-        </referenceContainer>
-    </body>
-</page>
