--- app/code/Gateway/view/frontend/templates/redirect.phtml
@@ -1,1 +1,0 @@
-<h1><?php echo "test"; //TODO post to ipgpay ?></h1>
