--- app/code/Gateway/view/frontend/web/js/action/set-payment-method.js
@@ -1,53 +1,0 @@
-define(
-    [
-        'jquery',
-        'Magento_Checkout/js/model/quote',
-        'Magento_Checkout/js/model/url-builder',
-        'mage/storage',
-        'Magento_Checkout/js/model/error-processor',
-        'Magento_Customer/js/model/customer',
-        'Magento_Checkout/js/model/full-screen-loader'
-    ],
-    function ($, quote, urlBuilder, storage, errorProcessor, customer, fullScreenLoader) {
-        'use strict';
-
-        return function (messageContainer) {
-            var serviceUrl,
-                payload,
-                paymentData = quote.paymentMethod();
-
-            /**
-             * Checkout for guest and registered customer.
-             */
-            if (!customer.isLoggedIn()) {
-                serviceUrl = urlBuilder.createUrl('/guest-carts/:cartId/selected-payment-method', {
-                    cartId: quote.getQuoteId()
-                });
-                payload = {
-                    cartId: quote.getQuoteId(),
-                    method: paymentData
-                };
-            } else {
-                serviceUrl = urlBuilder.createUrl('/carts/mine/selected-payment-method', {});
-                payload = {
-                    cartId: quote.getQuoteId(),
-                    method: paymentData
-                };
-            }
-            fullScreenLoader.startLoader();
-
-            return storage.put(
-                serviceUrl, JSON.stringify(payload)
-            ).done(
-                function () {
-                    $.mage.redirect(window.checkoutConfig.payment.ipgpay_gateway.redirectUrl);
-                }
-            ).fail(
-                function (response) {
-                    errorProcessor.process(response, messageContainer);
-                    fullScreenLoader.stopLoader();
-                }
-            );
-        };
-    }
-);
