+++ app/code/IPGPAY/Gateway/Api/Exceptions/CommunicationException.php
