+++ app/code/IPGPAY/Gateway/Api/Response/Declined.php
