+++ app/code/IPGPAY/Gateway/Api/Response/Error.php
