+++ app/code/IPGPAY/Gateway/Api/Response/ResponseAbstract.php
