+++ app/code/IPGPAY/Gateway/view/frontend/layout/ipgpay_redirect_index.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0"?>
+<!--
+/**
+ * Copyright © 2015 Magento. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+-->
+<page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" layout="1column" xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
+    <head>
+        <link src="IPGPAY_Gateway::js/view/payment/ipgpay_redirect.js" type='text/javascript'/>
+    </head>
+    <body>
+        <referenceContainer name="content">
+            <block class="IPGPAY\Gateway\Block\Redirect" name="redirect" template="IPGPAY_Gateway::redirect.phtml" cacheable="false"/>
+        </referenceContainer>
+    </body>
+</page>
