+++ app/code/IPGPAY/Gateway/view/frontend/templates/redirect.phtml
@@ -1,0 +1,6 @@
+<div><?php echo $this->escapeHtml($this->getRedirectMessage()); ?></div>
+<fieldset class="fieldset redirect" style="display:none;" id="payment_form_ipgpay">
+    <form id="postforwardform" method="post" action="<?php echo $this->escapeHtml($this->getPaymentFormUrl()); ?>" accept-charset="UTF-8">
+        <?php echo $this->getPaymentFormFields(); ?>
+    </form>
+</fieldset>
