+++ app/code/IPGPAY/Gateway/view/frontend/web/js/view/payment/ipgpay_gateway.js
