+++ app/code/IPGPAY/Gateway/view/frontend/web/js/view/payment/ipgpay_redirect.js
@@ -1,0 +1,8 @@
+require(['jquery'], function($){
+    $(document).ready(function(){
+        setTimeout(function() {
+            $("#postforwardform").submit();
+        }, 0);
+    });
+});
+
