+++ app/code/IPGPAY/Gateway/view/frontend/web/template/payment/ipgpay_gateway.html
